AP-7589: load workflow with correct URI to avoid key matching problems in ProjectWorkflowMap
If we convert the path directly to URI this will add three slashes after the scheme to the resulting URI. However, the value
will be accessed with only one slash after the scheme. As the ProjectWorkflowMap works on the String representation of URI
this will result in non matching keys. If we convert the path to file before converting to URI this will lead to the correct URI.

AP-7589 (During execution via Call Local Workflow nodes, the annotators at the workflows are not shown)