AP-7737 : Add async update of collapsible panels and add loading bar
When a local workflow is called and we tried to open the dialog while this wokflow was executing, the AP froze due to the main thread waiting to acquire the lock from the executing workflow. Therefore, the wokflow will be loaded async in a separate SwingWorker and an infinite loading bar will be shown as long as the lock is blocking.

AP-7737 (During execution via Call Local Workflow, opening the dialog will freeze the AP)