AP-7737 Add null check
If a new Call Local Workflow node is created the doInBackgroundWithContext method will
return null because the workflow path is empty. Then an NPE will be thrown in doneWithContext
and the Dialog will not be updated correctly.

AP-7737 (During execution via Call Local Workflow, opening the dialog
will freeze the AP)